+++ config.clj
@@ -1,5 +1,5 @@
 {
- :version "0.10.0"
+ :version "0.11.0"
  :clojure-version "1.4.0"
  :speclj-version "2.3.1"
 }