+++ lein-joodo/resources/joodo/kuzushi/templates/config/env.clj
@@ -1,7 +1,3 @@
-(use 'joodo.env)
-
-(def environment {
+(alter-env! assoc
   :joodo-env "!-ENV-!"
-  })
-
-(swap! *env* merge environment)
+  )